Feat: controller 90%완성
:+1:  aop , enum class 수정 , 파일 업로드 완료해야함

+++ back/src/main/java/com/routediary/control/AdminController.java
@@ -1,7 +1,10 @@
 package com.routediary.control;
 
+import java.util.List;
 import java.util.Optional;
 import javax.servlet.http.HttpSession;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -9,14 +12,22 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.multipart.MultipartFile;
 import com.routediary.dto.Admin;
 import com.routediary.dto.Diary;
+import com.routediary.dto.Notice;
 import com.routediary.dto.PageBean;
 import com.routediary.dto.ResultBean;
+import com.routediary.exception.AddException;
 import com.routediary.exception.FindException;
+import com.routediary.exception.ModifyException;
 import com.routediary.exception.RemoveException;
 import com.routediary.service.AdminServiceImpl;
 import com.routediary.service.NoticeServiceImpl;
@@ -28,6 +39,12 @@ public class AdminController {
   private AdminServiceImpl adminService;
   @Autowired
   private NoticeServiceImpl noticeService;
+  private Logger logger = LoggerFactory.getLogger(getClass());
+  
+  @GetMapping(value="write")
+  public String write() {
+    return "write";
+  }
 
   // -------로그인 관련 START
   @GetMapping(value = {"login"})
@@ -74,8 +91,7 @@ public class AdminController {
   // --------로그인 관련 END
   // --------다이어리 관련 START
   @GetMapping(value = {"diary/list", "diary/list/{currentPageOpt}"})
-  public ResultBean<PageBean<Diary>> showDiaryBoard(
-      @PathVariable Optional<Integer> currentPageOpt,
+  public ResultBean<PageBean<Diary>> showDiaryBoard(@PathVariable Optional<Integer> currentPageOpt,
       @RequestParam(name = "order", required = false, defaultValue = "2") int order)
       throws FindException {
     ResultBean<PageBean<Diary>> resultBean = new ResultBean<PageBean<Diary>>();
@@ -94,14 +110,15 @@ public class AdminController {
 
   }
 
-  @GetMapping(value = {"{diary}"})
-  public ResultBean<Diary> showDiary(@RequestParam(name="diaryNo" ,required=true)Optional<Integer> diaryNoOpt) throws FindException {
+  @GetMapping(value = {"diary/{diaryNoOpt}"})
+  public ResultBean<Diary> showDiary(@PathVariable Optional<Integer> diaryNoOpt)
+      throws FindException {
     ResultBean<Diary> resultBean = new ResultBean<Diary>();
     int diaryNo;
-    if(diaryNoOpt.isPresent()) {
+    if (diaryNoOpt.isPresent()) {
       diaryNo = diaryNoOpt.get();
-    }else {
-     throw new FindException("다이어리가 없습니다"); 
+    } else {
+      throw new FindException("다이어리가 없습니다");
     }
     Diary diary = adminService.showDiary(diaryNo);
     resultBean.setStatus(1);
@@ -110,11 +127,73 @@ public class AdminController {
     return resultBean;
 
   }
-  @DeleteMapping(value = {"{diary}"} , produces= MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<?> removeDiary(int diaryNo) throws RemoveException{
+
+  @DeleteMapping(value = {"diary/{diaryNo}"}, produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<String> removeDiary(@PathVariable int diaryNo, HttpSession session)
+      throws RemoveException {
+    String adminId = (String) session.getAttribute("loginInfo");
+    if (adminId == null) {
+      return new ResponseEntity<String>("관리자가 아닙니다", HttpStatus.BAD_REQUEST);
+    }
     adminService.removeDiary(diaryNo);
-    return null;
+    return new ResponseEntity<String>("삭제되었습니다", HttpStatus.OK);
+  }
+
+  @DeleteMapping(value = {"diary/{diaryNo}/comment"}, produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<?> removeComment(@PathVariable int diaryNo,
+      @RequestParam(name = "commentNo", required = true) int commentNo, HttpSession session)
+      throws RemoveException {
+    String adminId = (String) session.getAttribute("loginInfo");
+    if (adminId == null) {
+      return new ResponseEntity<String>("관리자가 아닙니다", HttpStatus.BAD_REQUEST);
+    }
+    adminService.removeComment(diaryNo, commentNo);
+    return new ResponseEntity<String>("댓글이 삭제되었습니다", HttpStatus.OK);
+
   }
 
   // --------다이어리 관련 END
+  // --------공지사항 관련 START
+  @PutMapping(value = {"notice/{noticeNo}"}, produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<?> modifyNotice(@PathVariable int noticeNo, @RequestBody Notice notice,
+      HttpSession session) throws ModifyException {
+
+    String adminId = (String) session.getAttribute("loginInfo");
+    if (notice.getNoticeTitle() == null || notice.getNoticeTitle().equals("")
+        || notice.getNoticeContent() == null || notice.getNoticeContent().equals("")) {
+      return new ResponseEntity<>("공지사항이 없습니다", HttpStatus.BAD_REQUEST);
+    } else if (adminId == null) {
+      return new ResponseEntity<>("관리자가 아닙니다", HttpStatus.BAD_REQUEST);
+    }
+    adminService.modifyNotice(notice);
+    return new ResponseEntity<>("공지사항이 수정되었습니다", HttpStatus.OK);
+
+  }
+  @DeleteMapping(value = {"notice/{noticeNo}"},produces= MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<?> removeNotice(@PathVariable @RequestParam(name="noticeNo",required=true) int noticeNo,HttpSession session) throws RemoveException{
+    
+    String adminId = (String) session.getAttribute("loginInfo");
+    if(adminId == null) {
+      return new ResponseEntity<>("관리자가 아닙니다", HttpStatus.BAD_REQUEST);
+    }
+    adminService.removeNotice(noticeNo);
+    return new ResponseEntity<>("공지사항이 삭제되었습니다", HttpStatus.OK);
+  }
+  @PostMapping(value= "notice/write", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<?> writeNotice(@RequestPart(value="files",required=false) List<MultipartFile> imgfiles,
+                                       @RequestPart(required=false) MultipartFile imgfile,
+                                       @RequestBody Notice notice,HttpSession session,
+                                       @RequestPart(required=false) String greeting) throws AddException{
+    logger.info(greeting);
+    String adminId = (String) session.getAttribute("loginInfo");
+    
+    notice.setAdminId(adminId);
+    adminService.writeNotice(notice);
+    return new ResponseEntity<>("공지사항이 작성되었습니다", HttpStatus.OK);
+    
+    // ---이미지 업로드 START
+//    /Users/minseong/Desktop
+    // ---이미지 업로드 END
+  }
+  // --------공지사항 관련 END
 }
